run button and export button work on all functions

+++ functions/alphashape.py
@@ -50,7 +50,7 @@ class Function(FunctionBase):
                     FROM node
                     JOIN
                     (SELECT * FROM pgr_drivingDistance(''
-                        SELECT %(id)s AS id,
+                        SELECT %(id)s::int4 AS id,
                             %(source)s::int4 AS source,
                             %(target)s::int4 AS target,
                             %(cost)s::float8 AS cost%(reverse_cost)s
